refactor : (Gym, GymLike) Controller, Service 통합
