+++ orury-client/src/test/java/org/orury/client/config/ControllerTest.java
@@ -1,6 +1,8 @@
 package org.orury.client.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.orury.client.auth.controller.AuthController;
 import org.orury.client.auth.jwt.JwtTokenProvider;
 import org.orury.client.auth.service.AuthService;
@@ -10,8 +12,6 @@ import org.orury.client.comment.controller.CommentLikeController;
 import org.orury.client.comment.service.CommentLikeService;
 import org.orury.client.comment.service.CommentService;
 import org.orury.client.gym.controller.GymController;
-import org.orury.client.gym.controller.GymLikeController;
-import org.orury.client.gym.service.GymLikeService;
 import org.orury.client.gym.service.GymService;
 import org.orury.client.post.controller.PostController;
 import org.orury.client.post.controller.PostLikeController;
@@ -24,8 +24,6 @@ import org.orury.client.review.service.ReviewService;
 import org.orury.client.user.controller.UserController;
 import org.orury.client.user.service.UserService;
 import org.orury.common.config.SlackMessage;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -40,7 +38,6 @@ import org.springframework.test.web.servlet.MockMvc;
         CommentController.class,
         CommentLikeController.class,
         GymController.class,
-        GymLikeController.class,
         PostController.class,
         PostLikeController.class,
         ReviewController.class,
@@ -66,9 +63,6 @@ public abstract class ControllerTest {
 
     @MockBean
     protected GymService gymService;
-
-    @MockBean
-    protected GymLikeService gymLikeService;
 
     @MockBean
     protected JwtTokenProvider jwtTokenProvider;
